requirements.txt
requests==2.29.0
beautifulsoup4
streamlit
numpy
Pillow
flask
tensorflow~=2.7.0
protobuf==3.20